[breadboard-ui] Support defaults for LLM Content in Schema Editor
